abstract out
